SW-7584 Add searchCount endpoint to search service (#3534)
In order to support BE pagination, we'll need to have a total row count to know how many pages and rows there are. Rather than adding this to the regular search api (and thus probably over-query this when it's not needed), add a separate endpoint to support this. 

The next PR will limit what is queried.