SW-7606 Differentiate photos from original observation (#3558)
Add a flag to observation photos to indicate whether a given photo was uploaded as
part of the original observation data or added later.

Currently, the flag is always true since there isn't yet any way to upload photos
after the fact, but future changes will add an alternate upload endpoint for
additional photo uploads.

Include the flag in the photo data that's returned as part of observation results
so the originality of photos can be indicated in the web app UI.