SW-7600 Indicate videos in observation results (#3564)
Add a new `mediaKind` property to the observation results photo payload to
indicate whether a given media file is a photo or a video.

Rename the payload class to replace `Photo` with `Media`. This affects the payload
name in the OpenAPI schema, but we aren't looking payload types up by name in our
client code, so this shouldn't require client changes.

Add a `media` property to the monitoring plot results payload alongside the
existing `photos` property (with identical contents) to allow the frontend code
to switch to the more accurate name at leisure.