SW-7250 Add dead plants to totalPlants per species in observation results (#3578)
The property `totalPlants` has a different definition depending on if it's per species or for a whole plot/subzone/zone/site. The FE currently calculates this in some cases but not all. Make `totalPlants` include live + existing + dead when calculating per species.

Note that this also covers SW-7598 which is a duplicate of SW-7250 .