SW-7684 Add endpoint for species required for t0 data (#3590)
The previous `getSpeciesWithdrawnToPlantingSite` endpoint only included species that were withdrawn to plots and their existing densities (for pre-population in the FE). We want to also know any species that were recorded in observations but not withdrawn to the plot's subzone.
Create a new endpoint that combines both lists of species. Use the same basic format for the return but make the density optional, such that null densities should be treated as species still required for t0 data, even though they don't have withdrawal densities to pre-populate the FE with.

The existing endpoint `getSpeciesWithdrawnToPlantingSite` will get removed in https://terraformation.atlassian.net/browse/SW-7688.