SW-7685 Add later-observed species to t0 densities when assigning observation (#3587)
When assigning an observation as t0 for a plot, species that were withdrawn but not recorded in the observation are set to 0 for their t0 density. We want to also include species that were recorded in later observations, regardless of if they were withdrawn or not. Add these species to the "zero list".