SW-7720 Use InMemoryFileStore in ThumbnailStoreTest
`ThumbnailStoreTest` was written before `InMemoryFileStore` existed, so it used
MockK to simulate a file store. Update it to use `InMemoryFileStore`.

Converting it exposed a minor bug in `ThumbnailStore` that was unlikely to have
caused any issues in production: it was writing the wrong file length to the
thumbnails table when the thumbnail file already existed in the file store.