SW-7720 Return full-sized published activity JPEGs (#3615)
If a published activity had a JPEG image, and someone had previously requested a
scaled-down thumbnail version, subsequent requests to fetch the full-sized
version would return the thumbnail instead because we assumed that the largest
existing thumbnail was always the full-sized version of the image. That's true
for HEIC images but not necessarily true for JPEGs.

Track whether or not each thumbnail is the same size as the original image, and
when a full-sized JPEG is requested, search for a full-sized thumbnail rather
than blindly returning the largest one.