SW-7694 All Set check should check all definitions of plot permanence (#3620)
When checking if all permanent plots have t0 data set, change the definition to include all of the following (previously was just numbers 1 & 2):

1. currently permanent
1. have a completed or abandoned observation
1. within that observation the plot was also permanent

This helps ensure that if a plot was previously temporary, had an observation, then was changed to permanent via a site geometry change, it will not be included incorrectly in the permanent plot list until it's had a completed observation.

Do the same thing for temporary plots in reverse.