SW-7607 Add event log query endpoint (#3595)
`POST /api/v1/events/list` will search the event log for events related to the IDs
specified in the request body, optionally limited to specific kinds of entities.

In this initial change, only observation media events are supported; others will
be added later. Example payload to query the media events for a particular
observation of a particular monitoring plot:

    {
        "organizationId": 1,
        "observationId": 2,
        "monitoringPlotId": 3,
        "subjects": ["ObservationPlotMedia"]
    }

Events are represented using "subjects" and "actions," where the list of actions
is currently limited to create, delete, and field-update. These always use the
same payload format no matter what the event is. The subject's structure varies
depending on the type of entity, but always includes long and short text
descriptions which UI code can use to display line items for events it doesn't
recognize or that don't need custom rendering logic.

In this initial change, the field names in the field-update action aren't
localized; that'll come later. (I say "field names" but there's only one field
so far: the "caption" field on observation photos/videos.)