Use fields-updated format in OrganizationRenamedEvent (#3628)
`OrganizationRenamedEventV1` was added before `FieldsUpdatedPersistentEvent` was
defined, so it didn't follow that interface's conventions. Introduce a new version
that follows the more recent pattern.

This is the first time we've added upgrade logic for an actual event, and it
exposed the fact that upgrades need to be able to fetch earlier events from the
log without worrying about whether the earlier events are themselves out of date.
`EventUpgradeUtils` can now call `EventLogStore` query methods, and the current
event's ID is passed to `UpgradableEvent.toNextVersion()` to allow it to limit
its searches to prior events.