Add test for localizable update event field names (#3639)
"Fields updated" action payloads return localized field names, which means the
field names need to be in the strings file.

For classes that implement `FieldsUpdatedPersistentEvent` and follow the typical
pattern of using `changedFrom` and `changedTo` properties with nested classes
for the field lists, we can sanity check that there are strings for all the
properties. This won't be foolproof but should catch cases where someone follows
the typical pattern but forgets to add strings to `Messages_en.properties`.