SW-7673 Added planting site histories to search (#3635)
These search tables are exposed for planting site map to get site histories with granular control of zones/subzones and geometry. For example, for planting site maps, we can search for a list of `plantingSiteHistoryId` and `createdTime` for the slider, without fetching full geometry down to subzone level. Once the user selects a version, and the desired level of site geometry, we can use this search API or the planting site history API to fetch the full geometry.