Support distinct null and absent payload properties (#3646)
To support PATCH-style API semantics, we need to be able to distinguish between
a payload property being absent from the incoming JSON (in which case the existing
value should be left as is) and the property being present with a value of `null`
(in which case the client is asking us to remove the existing value).

Jackson supports using the Java `Optional` class to distinguish those two cases,
but it's a little cumbersome out of the box; add an extension method to simplify
using it for PATCH-style operations.

The distinction between "not required" and "can be null" needs to be reflected
in the OpenAPI schema as well. We could do it by adding `@Schema(nullable = true)`
to payload properties, but that'll add a lot of noise, so instead add logic to
automatically mark all `Optional` payload properties as allowing null values.