Upgrade to Spring Boot 4
Upgrade Spring Boot to the latest release. This requires moving to newer versions
(or alternate implementations) of several dependencies.

Spring Boot 4 reorganizes some of Spring's classes, so we need to use the updated
names/locations. In addition, it uses Jackson 3 instead of Jackson 2; Jackson 3
uses a different package prefix for most of its classes, so our imports need to
be adjusted accordingly, and it uses a different scheme for programmatic
configuration.