Let fields-updated events localize their values (#3662)
The list of changed values returned by event classes that implement
`FieldsUpdatedPersistentEvent` are supposed to be localized, but there was
no good way for those classes to look up localized values other than enum
display names. Pass the `Messages` object to the list-changes function
so it can look up any other localized text it needs.

This change just adds the capability. None of the existing event classes
use it, but upcoming ones will.