SW-7695 Move recorded tree update to PATCH endpoint
Initially we were going to add separate API endpoints for updating each of the
kinds of observation data we need to be able to edit, but we switched to using
a single endpoint that can accept a list of changes. Add support for editing
recorded tree descriptions to the new endpoint and remove the old one, which
nothing is calling yet on the client side.