SW-7788 Fix NPE when fetching old plot details (#3687)
Exporting assigned plots to a GPX file for an old observation failed if there was
a map edit that caused one of the observation's plots to no longer lie within the
site boundary. The export would also have used the wrong subzone and zone names
if a map edit caused a plot to move to a different subzone. Update the code to
look up the subzone and zone names as of the time of the observation, and don't
return a subzone ID if the plot is no longer in the site.