SW-7725 Utilize correct t0 when a plot changes from perm to temp or vice versa (#3661)
When a plot changes from temp to permanent or permanent to temp, the observation results will use the incorrect t0 densities if there was an observation on the plot before the change. 

Additionally, the survival rate can be incorrect on a later observation, as calculated in the `completeObservation` method.

Fix these.

Also fix when a plot is removed from a subzone.