Make generic exception log message more informative
If a controller throws an exception that isn't one of the ones we have
special-case handling for, we log a generic "exception thrown" error message.
From the message alone, you can't tell what the error was; you have to go
look at the attached stack trace.

Update the logged error to include the exception name and its message.