SW-7810 Fix survival rates when unrecorded species in later observations (#3704)
When a species (with a t0 density) is recorded in an observation, and then not recorded in a later observation but other species in the plot are, the survival rate is incorrect (specifically, that species' t0 density is not being included in the survival rate denominator). Fix the full outer join on `plot_t0_densities` to correctly include all relevant rows.