SW-7807 Deprecate mortality rate system metric
We no longer want to calculate mortality rates when generating reports, but we
want to preserve the mortality rates on reports that have already been published
since the rates were part of their original contents.

Introduce the ability to mark a system metric as deprecated. Deprecated metrics
aren't calculated from Terraware data, but are still allowed to be present on
existing reports.