SW-7850 Add survivalRateIncludesTempPlots to planting site payload (#3724)
The flag for `survivalRateIncludesTempPlots` was not included in the planting site payload, causing it to get overwritten whenever an edit was made to the planting site. This is because the updatePlantingSite endpoint assumed that the content of the update payload contained what the planting site should be, and thus set the flag to false since it wasn't included. 
Add the flag to the payload.
(also correct an alphabetical typo)