SW-7856 Update to new logic for project level survival rates in reports (#3732)
The logic for handling multiple planting sites for a survival rate had not yet been implemented for reports. Implement this using the `ObservationResultsStore` in order to decrease duplication and because the logic for checking multiple planting sites' survival rates is even more complex than what we have for reports currently.