SW-7827 Rename variables in Observation-related files for Zone -> Strata/Stratum (#3718)
Update variables (and comments) in all files related to Observations, such as `ObservationStore` and `ObservationResultsStore`. This was taken to the extent such that there are no more references to zones or subzones in these files, which means that some files unrelated to observations were updated.

This ended up being larger than expected, and can be broken up into smaller PRs to make it easier to review if preferred.