SW-7892 Support tiny substrata (#3756)
We round geometry areas to one decimal place, which means that the area of a
substratum of less than 0.05 hectares gets rounded down to zero. This causes
planting site creation to fail because substratum areas must be greater than
zero.

Update the rounding logic so that it uses as many digits as needed for very
small areas.

This only applies to substrata; there is a minimum size requirement for strata
(they have to be big enough to hold monitoring plots) which is still enforced.