SW-7783 Add DSL for observation test scenarios
In preparation for testing the ability to edit observation results, add a little
DSL that allows tests to specify test scenarios in a structured, human-friendly
way instead of requiring tests to come with lots of little CSV files.

To demonstrate how the DSL is used, convert the existing "survival rate is
updated when t0 density changes" test case. Additional tests will be converted
in later changes.