SW-7857 Use web app's internal role name translations
The Spanish and French translations for project internal role names didn't match
between terraware-web and this repo, so when the frontend tried to use a role as a
search filter, it didn't work because the search string wasn't recognized as a
valid enum value. Update the translations to match the ones in terraware-web.