SW-7828 Rename zone->strata in search sublists (#3773)
Add sublists with updated names for all search tables. 
The mobile app uses the old names in `plantingSites` so those will have to stick around awhile to support backwards compatibility with old versions, but the rest of the old sublist names can be removed in a later PR once this is merged and the FE is updated as well.