Test disjoint substratum observations with map edit (#3781)
Add a test case to verify (and document) how survival rates are calculated when
there are geometry changes that affect substrata that aren't observed in later
observations.

This is also a trial run of specifying the expected values in a more
self-documenting way; it includes a new null-safe wrapper for `Map` that allows
us to do null-safe array-style gets without having to litter the code with `!!`
modifiers.

The test uncovered a bug in the test DSL's site editing logic; it wasn't
maintaining consistent stable IDs when substrata moved between strata.